+++ kraken/lib/src/dom/element.dart
@@ -125,6 +125,24 @@ class Element extends Node
 
   Size get viewportSize => elementManager.viewport.viewportSize;
 
+  /// Whether should create repaintBoundary for this element when style changed
+  bool get shouldConvertToRepaintBoundary {
+    // Following cases should always convert to repaint boundary for performance consideration
+    // Multiframe image
+    bool isMultiframeImage = this is ImageElement && (this as ImageElement).isMultiframe;
+    // Intrinsic element such as Canvas
+    bool isSetRepaintSelf = repaintSelf;
+    // Scrolling box
+    bool isScrollingBox = scrollingContentLayoutBox != null;
+    // Transform element
+    bool hasTransform = renderBoxModel?.renderStyle?.transform != null;
+    // Fixed element
+    bool isPositionedFixed = renderBoxModel?.renderStyle?.position == CSSPositionType.fixed;
+
+    return isMultiframeImage || isScrollingBox ||
+      isSetRepaintSelf || hasTransform || isPositionedFixed;
+  }
+
   Element(int targetId, this.nativeElementPtr, ElementManager elementManager,
       {this.tagName,
         this.defaultStyle = const <String, dynamic>{},
@@ -137,7 +155,7 @@ class Element extends Node
       : assert(targetId != null),
         assert(tagName != null),
         _isIntrinsicBox = isIntrinsicBox,
-        defaultDisplay = defaultStyle.containsKey(DISPLAY) ? defaultStyle[DISPLAY] : BLOCK,
+        defaultDisplay = defaultStyle.containsKey(DISPLAY) ? defaultStyle[DISPLAY] : INLINE,
         super(NodeType.ELEMENT_NODE, targetId, nativeElementPtr.ref.nativeNode, elementManager, tagName) {
     style = CSSStyleDeclaration(this);
 
@@ -396,19 +414,15 @@ class Element extends Node
     }
   }
 
-  /// Convert RenderIntrinsic to non repaint boundary
-  void _convertToNonRepaint() {
-    // Multiframe image should always convert to repaint boundary for scroll performance
-    if (this is ImageElement && (this as ImageElement).isMultiframe) {
-      return;
-    }
+  /// Convert renderBoxModel to non repaint boundary
+  void convertToNonRepaintBoundary() {
     if (renderBoxModel != null && renderBoxModel.isRepaintBoundary) {
       toggleRepaintSelf(repaintSelf: false);
     }
   }
 
-  /// Convert RenderIntrinsic to repaint boundary
-  void _convertToRepaint() {
+  /// Convert renderBoxModel to repaint boundary
+  void convertToRepaintBoundary() {
     if (renderBoxModel != null && !renderBoxModel.isRepaintBoundary) {
       toggleRepaintSelf(repaintSelf: true);
     }
@@ -449,15 +463,37 @@ class Element extends Node
 
     // Move element according to position when it's already attached to render tree.
     if (isRendererAttached) {
-      RenderObject prev = previousSibling?.renderer;
-      detach();
-      attachTo(parent, after: prev);
+      RenderObject prev = (renderer.parentData as ContainerBoxParentData).previousSibling;
+      // It needs to find the previous sibling of the previous sibling if the placeholder of
+      // positioned element exists and follows renderObject at the same time, eg.
+      // <div style="position: relative"><div style="postion: absolute" /></div>
+      if (prev == renderBoxModel) {
+        prev = (renderBoxModel.parentData as ContainerBoxParentData).previousSibling;
+      }
+
+      // Remove placeholder of positioned element.
+      RenderPositionHolder renderPositionHolder = renderBoxModel.renderPositionHolder;
+      if (renderPositionHolder != null) {
+        ContainerRenderObjectMixin parent = renderPositionHolder.parent;
+        if (parent != null) {
+          parent.remove(renderPositionHolder);
+          renderBoxModel.renderPositionHolder = null;
+        }
+      }
+      // Remove renderBoxModel from original parent and append to its containing block
+      RenderObject parentRenderBoxModel = renderBoxModel.parent;
+      if (parentRenderBoxModel is ContainerRenderObjectMixin) {
+        parentRenderBoxModel.remove(renderBoxModel);
+      } else if (parentRenderBoxModel is RenderProxyBox) {
+        parentRenderBoxModel.child = null;
+      }
+      parent.addChildRenderObject(this, after: prev);
     }
 
-    if (currentPosition == CSSPositionType.fixed) {
-      _convertToRepaint();
+    if (shouldConvertToRepaintBoundary) {
+      convertToRepaintBoundary();
     } else {
-      _convertToNonRepaint();
+      convertToNonRepaintBoundary();
     }
 
     // Add fixed children after convert to repaint boundary renderObject
@@ -576,15 +612,15 @@ class Element extends Node
     }
 
     /// Recalculate gradient after node attached when gradient length cannot be obtained from style
-    if (renderBoxModel.recalGradient) {
+    if (renderBoxModel.shouldRecalGradient) {
       String backgroundImage = style[BACKGROUND_IMAGE];
       renderBoxModel.renderStyle.updateBox(BACKGROUND_IMAGE, backgroundImage, backgroundImage);
-      renderBoxModel.recalGradient = false;
+      renderBoxModel.shouldRecalGradient = false;
     }
 
     /// Calculate font-size which is percentage when node attached
     /// where it can access the font-size of its parent element
-    if (renderBoxModel.parseFontSize) {
+    if (renderBoxModel.shouldLazyCalFontSize) {
       RenderStyle parentRenderStyle = parent.renderBoxModel.renderStyle;
       double parentFontSize = parentRenderStyle.fontSize ?? CSSText.DEFAULT_FONT_SIZE;
       double parsedFontSize = parentFontSize * CSSLength.parsePercentage(style[FONT_SIZE]);
@@ -594,12 +630,12 @@ class Element extends Node
           node.updateTextStyle();
         }
       }
-      renderBoxModel.parseFontSize = false;
+      renderBoxModel.shouldLazyCalFontSize = false;
     }
 
     /// Calculate line-height which is percentage when node attached
     /// where it can access the font-size of its own element
-    if (renderBoxModel.parseLineHeight) {
+    if (renderBoxModel.shouldLazyCalLineHeight) {
       RenderStyle renderStyle = renderBoxModel.renderStyle;
       double fontSize = renderStyle.fontSize ?? CSSText.DEFAULT_FONT_SIZE;
       double parsedLineHeight = fontSize * CSSLength.parsePercentage(style[LINE_HEIGHT]);
@@ -609,7 +645,7 @@ class Element extends Node
           node.updateTextStyle();
         }
       }
-      renderBoxModel.parseLineHeight = false;
+      renderBoxModel.shouldLazyCalLineHeight = false;
     }
 
     RenderStyle renderStyle = renderBoxModel.renderStyle;
@@ -831,7 +867,7 @@ class Element extends Node
     return displayValue == INLINE;
   }
 
-  void _onStyleChanged(String property, String original, String present, bool inAnimation) {
+  void _onStyleChanged(String property, String original, String present) {
     switch (property) {
       case DISPLAY:
         _styleDisplayChangedListener(property, original, present);
@@ -1001,9 +1037,6 @@ class Element extends Node
   }
 
   void _styleDisplayChangedListener(String property, String original, String present) {
-    // Display change may case width/height doesn't works at all.
-    _styleSizeChangedListener(property, original, present);
-
     renderBoxModel.renderStyle.updateDisplay(present, this);
   }
 
@@ -1049,7 +1082,7 @@ class Element extends Node
   }
 
   void _styleOverflowChangedListener(String property, String original, String present) {
-    updateRenderOverflow(renderBoxModel, this, _scrollListener);
+    updateRenderOverflow(this, _scrollListener);
   }
 
   void _stylePaddingChangedListener(String property, String original, String present) {
@@ -1138,6 +1171,7 @@ class Element extends Node
     if (RenderStyle.isTransformTranslatePercentage(present)) return;
 
     Matrix4 matrix4 = CSSTransform.parseTransform(present, viewportSize);
+    // @FIXME support `none` value
     renderBoxModel.renderStyle.updateTransform(matrix4);
   }
 
@@ -1151,14 +1185,14 @@ class Element extends Node
     /// Percentage font-size should be resolved when node attached
     /// cause it needs to know its parents style
     if (property == FONT_SIZE && CSSLength.isPercentage(style[FONT_SIZE])) {
-      renderBoxModel.parseFontSize = true;
+      renderBoxModel.shouldLazyCalFontSize = true;
       return;
     }
 
     /// Percentage line-height should be resolved when node attached
     /// cause it needs to know other style in its own element
     if (property == LINE_HEIGHT && CSSLength.isPercentage(style[LINE_HEIGHT])) {
-      renderBoxModel.parseLineHeight = true;
+      renderBoxModel.shouldLazyCalLineHeight = true;
       return;
     }