+++ scripts/build_android_so.js
@@ -1,12 +1,27 @@
-require('./tasks');
-const { series } = require('gulp');
+const { paths } = require('./tasks');
+const { series, task } = require('gulp');
 const chalk = require('chalk');
+const { execSync } = require('child_process');
+const os = require('os');
+const path = require('path');
+const { copyFileSync } = require('fs');
+
+
+const buildTasks = [
+  'android-so-clean',
+  'compile-polyfill',
+  'build-android-kraken-lib'
+];
+
+if (os.platform() == 'win32') {
+  buildTasks.push(
+    'patch-windows-symbol-link-for-android'
+  );
+}
 
 // Run tasks
 series(
-  'sdk-clean',
-  'compile-polyfill',
-  'build-android-kraken-lib'
+  buildTasks
 )((err) => {
   if (err) {
     console.log(err);