+++ scripts/prepare_kraken_release_binary.js
@@ -1,14 +1,26 @@
 require('./tasks');
 const chalk = require('chalk');
 const { series } = require('gulp');
+const os = require('os');
 
-series(
+let buildTasks = [
   'sdk-clean',
   'compile-polyfill',
-  'build-darwin-kraken-lib',
-  'build-ios-kraken-lib',
   'build-android-kraken-lib',
-)((err) => {
+];
+
+if (os.platform() == 'win32') {
+  buildTasks.push(
+    'patch-windows-symbol-link-for-android'
+  );
+} else {
+  buildTasks.push(
+    'build-darwin-kraken-lib',
+    'build-ios-kraken-lib',
+  );
+}
+
+series(buildTasks)((err) => {
   if (err) {
     console.log(err);
   } else {