+++ sdk/.android/app/src/main/java/com/openkraken/kraken/host/MainActivity.java
@@ -1,6 +1,0 @@
-package com.openkraken.kraken.host;
-
-import io.flutter.embedding.android.FlutterActivity;
-
-public class MainActivity extends FlutterActivity {
-}